Add the website of governments around the world using French
I have tried to do my best to add websites of governments written in
French to offer a variety of starting points for the French language and
reflect the fact that French is an official language of many countries.

Two are missing:
* Guinée Equatorial https://www.guineaecuatorialpress.com/
* Comores: https://www.beit-salam.km/

There is no deep link to access to the French version

I'm a native speaker.